cart functionalties has been implemented

+++ app/src/main/java/com/giraffe/triplemapplication/model/cart/request/LineItem.kt
@@ -2,5 +2,5 @@ package com.giraffe.triplemapplication.model.cart.request
 
 data class LineItem(
     val quantity: Int,
-    val variant_id: Int
+    val variant_id: Long
 )
