+++ app/src/main/java/com/giraffe/triplemapplication/model/currency/Rates.kt
@@ -2,7 +2,7 @@ package com.giraffe.triplemapplication.model.currency
 
 data class Rates(
     val EGP: Double,
-    val EUR: Int,
+    val EUR: Double,
     val GBP: Double,
     val USD: Double
 )
